greenplum_test: make use of MustCreateCluster
Move MustCreateCluster() to greenplum/common_test.go and use it
consistently.
