+++ greenplum/start_or_stop_cluster_test.go
@@ -44,21 +44,6 @@ func init() {
 	)
 }
 
-// TODO: Consolidate with the same function in common_test.go in the
-//  hub package. This is tricky due to cycle imports and other issues.
-// MustCreateCluster creates a utils.Cluster and calls t.Fatalf() if there is
-// any error.
-func MustCreateCluster(t *testing.T, segs []SegConfig) *Cluster {
-	t.Helper()
-
-	cluster, err := NewCluster(segs)
-	if err != nil {
-		t.Fatalf("%+v", err)
-	}
-
-	return cluster
-}
-
 func TestStartOrStopCluster(t *testing.T) {
 	testhelper.SetupTestLogger() // initialize gplog
